Added command to fix the seed to a particular value. Kept the existing seed command the same, so macros should be backwards compatible.
Changed volID assignment to instead take a replacement pattern
to determine the volID to set based on the match. This casts
the replacement string to an int, so it should not break
any existing functionality -- if you use a constant int
for the second argument, it will just be cast to int appropriately
anyway.